[添加] 1、用户还书，通知关注这本书的用户去借书的邮件提示！

+++ BookManager/src/main/java/com/bookmanager/book/dto/BookListDTO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class BookListDTO {
     private Long id;
 
-    private Long isbn;
+    private Integer isbn;
 
     private String name;
