+++ BookManager/src/main/java/com/bookmanager/book/dto/RelationBookEmpDTO.java
@@ -7,6 +7,6 @@ import java.util.Date;
 @Data
 public class RelationBookEmpDTO {
     private Integer jobNumber; //用户编号
-    private Long isbn; //图书编号
+    private Integer isbn; //图书编号
     private Date currentTime;
 }
